Added linux build script

+++ src/main/java/io/webfolder/pdcurses4j/Window.java
@@ -13,8 +13,8 @@ public class Window {
     public static final short COLOR_MAGENTA = (COLOR_RED | COLOR_BLUE);
     public static final short COLOR_YELLOW = (COLOR_RED | COLOR_GREEN);
 
-    public static final long PDC_COLOR_SHIFT = 24;
-    public static final long A_COLOR = 0xff000000;
+    public static final int  PDC_COLOR_SHIFT = 24;
+    public static final int A_COLOR = 0xff000000;
 
     public static final int FALSE = 0;
     public static final int TRUE = 1;
@@ -24,32 +24,32 @@ public class Window {
 
     /*** Video attribute macros ***/
 
-    public static final long A_NORMAL = 0;
+    public static final int A_NORMAL = 0;
 
-    public static final long A_ALTCHARSET = 0x00010000;
-    public static final long A_RIGHT = 0x00020000;
-    public static final long A_LEFT = 0x00040000;
-    public static final long A_ITALIC = 0x00080000;
-    public static final long A_UNDERLINE = 0x00100000;
-    public static final long A_REVERSE = 0x00200000;
-    public static final long A_BLINK = 0x00400000;
-    public static final long A_BOLD = 0x00800000;
+    public static final int A_ALTCHARSET = 0x00010000;
+    public static final int A_RIGHT = 0x00020000;
+    public static final int A_LEFT = 0x00040000;
+    public static final int A_ITALIC = 0x00080000;
+    public static final int A_UNDERLINE = 0x00100000;
+    public static final int A_REVERSE = 0x00200000;
+    public static final int A_BLINK = 0x00400000;
+    public static final int A_BOLD = 0x00800000;
 
-    public static final long A_ATTRIBUTES = 0xffff0000;
-    public static final long A_CHARTEXT = 0x0000ffff;
+    public static final int A_ATTRIBUTES = 0xffff0000;
+    public static final int A_CHARTEXT = 0x0000ffff;
 
-    public static final long A_LEFTLINE = A_LEFT;
-    public static final long A_RIGHTLINE = A_RIGHT;
-    public static final long A_STANDOUT = (A_REVERSE | A_BOLD); /* X/Open */
+    public static final int A_LEFTLINE = A_LEFT;
+    public static final int A_RIGHTLINE = A_RIGHT;
+    public static final int A_STANDOUT = (A_REVERSE | A_BOLD); /* X/Open */
 
-    public static final long A_DIM = A_NORMAL;
-    public static final long A_INVIS = A_NORMAL;
-    public static final long A_PROTECT = A_NORMAL;
+    public static final int A_DIM = A_NORMAL;
+    public static final int A_INVIS = A_NORMAL;
+    public static final int A_PROTECT = A_NORMAL;
 
-    public static final long A_HORIZONTAL = A_NORMAL;
-    public static final long A_LOW = A_NORMAL;
-    public static final long A_TOP = A_NORMAL;
-    public static final long A_VERTICAL = A_NORMAL;
+    public static final int A_HORIZONTAL = A_NORMAL;
+    public static final int A_LOW = A_NORMAL;
+    public static final int A_TOP = A_NORMAL;
+    public static final int A_VERTICAL = A_NORMAL;
 
     @Deprecated
     public static final long CHR_MSK = A_CHARTEXT; /* Obsolete */
